changed data type from int to double

+++ src/main/java/godevenner/divexbackend/portfolio/dto/PortfolioEntryRequest.java
@@ -2,7 +2,7 @@ package godevenner.divexbackend.portfolio.dto;
 
 public record PortfolioEntryRequest(
     String ticker,
-    int stockPrice,
+    double stockPrice,
     int quantity,
 
     long portfolioId
